requirements.txt
pyaudio
pynput
numpy
faster-whisper
rumps
pyobjc
py2app
openai
python-dotenv
pyperclip